requirements.txt
django
djangorestframework
djangorestframework-simplejwt
drf-spectacular
django-cors-headers
python-dotenv
psycopg2-binary
